Provide IDESlave output with docs and type
Now, IDESlave provides the docstring overview and type of names that it
decorates. The module Idris.Output was introduced to remove some cyclic
dependency problems.
