+++ src/Idris/Error.hs
@@ -5,6 +5,7 @@ module Idris.Error where
 import Prelude hiding (catch)
 import Idris.AbsSyntax
 import Idris.Delaborate
+import Idris.Output
 
 import Idris.Core.TT
 import Idris.Core.Typecheck
